requirements.txt
streamlit
ultralytics
opencv-python
torch
torchvision
numpy
Pillow
